remove empty diagram elements
